Split event filter expression parser classes across multiple files
