+++ plaso/filters/expression_parser.py
@@ -92,41 +92,15 @@ filter is easy. Three basic filter implementations are given:
 
 from __future__ import unicode_literals
 
-import abc
 import binascii
-import calendar
 import codecs
-import datetime
 import logging
 import re
 
-from plaso.formatters import manager as formatters_manager
-from plaso.formatters import mediator as formatters_mediator
-from plaso.lib import definitions
+from plaso.filters import expressions
+from plaso.filters import filters
+from plaso.filters import value_expanders
 from plaso.lib import errors
-from plaso.lib import py2to3
-from plaso.lib import timelib
-
-
-def GetUnicodeString(value):
-  """Attempts to convert the argument to a Unicode string.
-
-  Args:
-    value (list|int|bytes|str): value to convert.
-
-  Returns:
-    str: string representation of the argument.
-  """
-  if isinstance(value, list):
-    value = [GetUnicodeString(item) for item in value]
-    return ''.join(value)
-
-  if isinstance(value, py2to3.INTEGER_TYPES):
-    value = '{0:d}'.format(value)
-
-  if not isinstance(value, py2to3.UNICODE_TYPE):
-    return codecs.decode(value, 'utf8', 'ignore')
-  return value
 
 
 class Token(object):
@@ -165,1085 +139,6 @@ class Token(object):
       self.actions = actions.split(',')
 
 
-class Expression(object):
-  """An event filter parser expression.
-
-  Attributes:
-    attribute (str): attribute or None if not set.
-    args (list[str]): arguments.
-    number_of_args (int): expected number of arguments.
-    operator (str): operator or None if not set.
-  """
-
-  # TODO: this currently needs to be a class attribute for objectfilter.
-  # See if this can be changed to an instance attribute.
-  attribute = None
-
-  def __init__(self):
-    """Initializes an event filter parser expression."""
-    super(Expression, self).__init__()
-    self.args = []
-    self.number_of_args = 1
-    self.operator = None
-
-  def AddArg(self, argument):
-    """Adds a new argument to this expression.
-
-    Args:
-       argument (str): argument to add.
-
-    Returns:
-      bool: True if the argument is the last argument, False otherwise.
-
-    Raises:
-      ParseError: If there are too many arguments.
-    """
-    self.args.append(argument)
-    if len(self.args) > self.number_of_args:
-      raise errors.ParseError('Too many arguments for this expression.')
-
-    elif len(self.args) == self.number_of_args:
-      return True
-
-    return False
-
-  @abc.abstractmethod
-  def Compile(self, filter_implementation):
-    """Given a filter implementation, compile this expression.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the binary expression.
-    """
-
-  def SetAttribute(self, attribute):
-    """Sets the attribute.
-
-    Args:
-      attribute (str): attribute, or None if not set.
-    """
-    self.attribute = attribute
-
-  def SetOperator(self, operator):
-    """Set the operator.
-
-    Args:
-      operator (str): operator, such as "and" or "&&", or None if not set.
-    """
-    self.operator = operator
-
-
-class BinaryExpression(Expression):
-  """An event filter parser expression which takes two other expressions."""
-
-  def __init__(self, operator='', part=None):
-    """Initializes an event filter parser binary expression.
-
-    Args:
-      operator (str): operator, such as "and" or "&&".
-      part (str): expression part.
-    """
-    super(BinaryExpression, self).__init__()
-    self.args = []
-    self.operator = operator
-
-    if part:
-      self.args.append(part)
-
-  def AddOperands(self, lhs, rhs):
-    """Adds an operand.
-
-    Args:
-      lhs (Expression): left hand side expression.
-      rhs (Expression): right hand side expression.
-
-    Raises:
-      ParseError: if either left hand side or right hand side expression
-          is not an instance of Expression.
-    """
-    if not isinstance(lhs, Expression):
-      raise errors.ParseError('Left hand side is not an expression')
-
-    if not isinstance(rhs, Expression):
-      raise errors.ParseError('Right hand side is not an expression')
-
-    self.args = [lhs, rhs]
-
-  def Compile(self, filter_implementation):
-    """Compiles the binary expression into a filter object.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the binary expression.
-    """
-    operator = self.operator.lower()
-    if operator in ('and', '&&'):
-      method = 'AndFilter'
-    elif operator in ('or', '||'):
-      method = 'OrFilter'
-    else:
-      raise errors.ParseError(
-          'Invalid binary operator {0:s}.'.format(operator))
-
-    args = [x.Compile(filter_implementation) for x in self.args]
-    return filter_implementation.FILTERS[method](arguments=args)
-
-
-class IdentityExpression(Expression):
-  """An event filter parser expression which always evaluates to True."""
-
-  def Compile(self, filter_implementation):
-    """Compiles the binary expression into a filter object.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the identity expression.
-    """
-    return filter_implementation.IdentityFilter()
-
-
-class Filter(object):
-  """Filter interface.
-
-  Attributes:
-    args (list[object]): arguments provided to the filter.
-    value_expander (ValueExpander): value expanded that is used to expand
-        arguments into comparable values.
-    value_expander_cls (type): value expander class.
-  """
-
-  def __init__(self, arguments=None, value_expander=None):
-    """Initializes a filter.
-
-    Implementations expanders are provided by subclassing ValueExpander.
-
-    Args:
-      arguments (Optional[object]): arguments.
-      value_expander (Optional[type]): class that is used to expand arguments
-          into comparable values.
-
-    Raises:
-      ValueError: If value expander is not a subclass of ValueExpander.
-    """
-    logging.debug('Adding {0!s}'.format(arguments))
-
-    if value_expander:
-      if not issubclass(value_expander, ValueExpander):
-        raise ValueError('{0:s} is not of type ValueExpander'.format(
-            type(value_expander)))
-
-    super(Filter, self).__init__()
-    self.args = arguments or []
-    self.value_expander = None
-    self.value_expander_cls = value_expander
-
-    if value_expander:
-      self.value_expander = value_expander()
-
-  @abc.abstractmethod
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-
-  def Filter(self, objects):
-    """Retrieves objects that match the filter.
-
-    Args:
-      objects (list[object]): objects to filter.
-
-    Returns:
-      list[object]: objects that match the filter.
-    """
-    return filter(self.Matches, objects)
-
-  def __str__(self):
-    """Retrieve a string representation of the filter.
-
-    Returns:
-      str: a string representation of the filter.
-    """
-    return '{0:s}({1:s})'.format(self.__class__.__name__, ', '.join([
-        str(argument) for argument in self.args]))
-
-
-class AndFilter(Filter):
-  """A filter that performs a boolean AND on the arguments.
-
-  Note that if no conditions are passed, all objects will pass.
-  """
-
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-    for child_filter in self.args:
-      if not child_filter.Matches(obj):
-        return False
-    return True
-
-
-class OrFilter(Filter):
-  """A filter that performs a boolean OR on the arguments.
-
-  Note that if no conditions are passed, all objects will pass.
-  """
-
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-    if not self.args:
-      return True
-
-    for child_filter in self.args:
-      if child_filter.Matches(obj):
-        return True
-    return False
-
-
-class Operator(Filter):
-  """Interface for filters that represent operators."""
-
-  @abc.abstractmethod
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-
-
-class IdentityFilter(Operator):
-  """A filter which always evaluates to True."""
-
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True to indicated the object matches the filter.
-    """
-    return True
-
-
-# TODO: remove, class is not used.
-class UnaryOperator(Operator):
-  """Interface for filters that represent unary operators."""
-
-  def __init__(self, operand, **kwargs):
-    """Initializes an unary operator.
-
-    Args:
-      operand (str): operand.
-
-    Raises:
-      InvalidNumberOfOperands: if the number of operands provided is not
-          supported.
-    """
-    if len(operand) != 1:
-      raise errors.InvalidNumberOfOperands((
-          '{0:s} only supports 1 operand, provided were {1:d} '
-          'operands.').format(self.__class__.__name__, len(operand)))
-
-    super(UnaryOperator, self).__init__(arguments=[operand], **kwargs)
-
-  @abc.abstractmethod
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-
-
-class BinaryOperator(Operator):
-  """Interface for binary operators.
-
-  Attributes:
-    left_operand (object): left hand operand.
-    right_operand (object): right hand operand.
-  """
-
-  def __init__(self, arguments=None, **kwargs):
-    """Initializes a binary operator.
-
-    Args:
-      arguments (Optional[object]): operands of the filter.
-
-    Raises:
-      InvalidNumberOfOperands: if the number of operands provided is not
-          supported.
-    """
-    if len(arguments) != 2:
-      raise errors.InvalidNumberOfOperands((
-          '{0:s} only supports 2 operands, provided were {1:d} '
-          'operands.').format(self.__class__.__name__, len(arguments)))
-
-    super(BinaryOperator, self).__init__(arguments=arguments, **kwargs)
-    self.left_operand = arguments[0]
-    self.right_operand = arguments[1]
-
-  @abc.abstractmethod
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-
-
-class GenericBinaryOperator(BinaryOperator):
-  """Shared functionality for common binary operators.
-
-  Attribute:
-    bool_value (bool): boolean value that represents the result of
-        the operation.
-  """
-
-  def __init__(self, arguments=None, **kwargs):
-    """Initializes a generic binary operator.
-
-    Args:
-      arguments (Optional[object]): operands of the filter.
-    """
-    super(GenericBinaryOperator, self).__init__(arguments=arguments, **kwargs)
-    self.bool_value = True
-
-  def FlipBool(self):
-    """Negates the value of the bool_value attribute."""
-    logging.debug('Negative matching.')
-    self.bool_value = not self.bool_value
-
-  @abc.abstractmethod
-  def Operation(self, x, y):
-    """Compares two values with the operator.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the values match according to the operator, False otherwise.
-    """
-
-  def Operate(self, values):
-    """Determines if one or more values match the filter.
-
-    Args:
-      values (list[object]): values to match against the filter.
-
-    Returns:
-      bool: True if one or more values match, False otherwise.
-    """
-    for val in values:
-      try:
-        if self.Operation(val, self.right_operand):
-          return True
-      except (TypeError, ValueError):
-        pass
-
-    return False
-
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-    key = self.left_operand
-    values = self.value_expander.Expand(obj, key)
-    values = list(values)
-    if values and self.Operate(values):
-      return self.bool_value
-    return not self.bool_value
-
-
-class Equals(GenericBinaryOperator):
-  """Equals (==) operator."""
-
-  def Operation(self, x, y):
-    """Compares if two values are equal.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the values are equal, False otherwise.
-    """
-    return x == y
-
-
-class NotEquals(GenericBinaryOperator):
-  """Not equals (!=) operator."""
-
-  def Operation(self, x, y):
-    """Compares if two values are not equal.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the values are not equal, False otherwise.
-    """
-    return x != y
-
-
-class Less(GenericBinaryOperator):
-  """Less than (<) operator."""
-
-  def Operation(self, x, y):
-    """Compares if the first value is less than the second.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value is less than the second, False otherwise.
-    """
-    return x < y
-
-
-class LessEqual(GenericBinaryOperator):
-  """Less than or equals (<=) operator."""
-
-  def Operation(self, x, y):
-    """Compares if the first value is less than or equals the second.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value is than or equals the second, False
-          otherwise.
-    """
-    return x <= y
-
-
-class Greater(GenericBinaryOperator):
-  """Greater than (>) operator."""
-
-  def Operation(self, x, y):
-    """Compares if the first value is greater than the second.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value is greater than the second, False otherwise.
-    """
-    return x > y
-
-
-class GreaterEqual(GenericBinaryOperator):
-  """Greater than or equals (>=) operator."""
-
-  def Operation(self, x, y):
-    """Compares if the first value is greater than or equals the second.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value is greater than or equals the second, False
-          otherwise.
-    """
-    return x >= y
-
-
-class Contains(GenericBinaryOperator):
-  """Operator to determine if a value contains another value."""
-
-  def Operation(self, x, y):
-    """Compares if the second value is part of the first.
-
-    Note that this method will do a case insensitive comparion if the first
-    value is a string.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the second value is part of the first, False otherwise.
-    """
-    if isinstance(x, py2to3.STRING_TYPES):
-      return y.lower() in x.lower()
-
-    return y in x
-
-
-# TODO: Change to an N-ary Operator?
-class InSet(GenericBinaryOperator):
-  """Operator to determine if a value is part of another value."""
-
-  def Operation(self, x, y):
-    """Compares if the first value is part of the second.
-
-    Note that this method will do a case insensitive string comparion if
-    the first value is a string.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value is part of the second, False otherwise.
-    """
-    if x in y:
-      return True
-
-    # x might be an iterable
-    # first we need to skip strings or we'll do silly things
-    # pylint: disable=consider-merging-isinstance
-    if isinstance(x, py2to3.STRING_TYPES) or isinstance(x, bytes):
-      return False
-
-    try:
-      for value in x:
-        if value not in y:
-          return False
-      return True
-    except TypeError:
-      return False
-
-
-# TODO: is GenericBinaryOperator the most suitable super class here?
-# Would BinaryOperator be a better fit?
-class Regexp(GenericBinaryOperator):
-  """Operator to determine if a value matches a regular expression.
-
-  Attributes:
-    compiled_re (???): compiled regular expression.
-  """
-
-  def __init__(self, arguments=None, **kwargs):
-    """Initializes a regular expression operator.
-
-    This operator uses case senstive comparision.
-
-    Args:
-      arguments (Optional[object]): operands of the filter.
-
-    Raises:
-      ValueError: if the regular expression is malformed.
-    """
-    super(Regexp, self).__init__(arguments=arguments, **kwargs)
-
-    # Note that right_operand is not necessarily a string.
-    logging.debug('Compiled: {0!s}'.format(self.right_operand))
-
-    try:
-      expression = GetUnicodeString(self.right_operand)
-      compiled_re = re.compile(expression, re.DOTALL)
-    except re.error:
-      raise ValueError('Regular expression "{0!s}" is malformed.'.format(
-          self.right_operand))
-
-    self.compiled_re = compiled_re
-
-  def Operation(self, x, y):
-    """Compares if the first value matches a regular expression.
-
-    Args:
-      x (object): first value.
-      y (object): second value.
-
-    Returns:
-      bool: True if the first value matches the regular expression, False
-          otherwise.
-    """
-    try:
-      string_value = GetUnicodeString(x)
-      if self.compiled_re.search(string_value):
-        return True
-    except TypeError:
-      pass
-
-    return False
-
-
-class RegexpInsensitive(Regexp):
-  """Operator to determine if a value matches a regular expression."""
-
-  def __init__(self, arguments=None, **kwargs):
-    """Initializes a regular expression operator.
-
-    This operator uses case insenstive comparision.
-
-    Args:
-      arguments (Optional[object]): operands of the filter.
-
-    Raises:
-      ValueError: if the regular expression is malformed.
-    """
-    super(RegexpInsensitive, self).__init__(arguments=arguments, **kwargs)
-
-    # Note that right_operand is not necessarily a string.
-    logging.debug('Compiled: {0!s}'.format(self.right_operand))
-
-    try:
-      expression = GetUnicodeString(self.right_operand)
-      compiled_re = re.compile(expression, re.I | re.DOTALL)
-    except re.error:
-      raise ValueError('Regular expression "{0!s}" is malformed.'.format(
-          self.right_operand))
-
-    self.compiled_re = compiled_re
-
-
-class Context(Operator):
-  """Restricts the child operators to a specific context within the object.
-
-  Solves the context problem. The context problem is the following:
-  Suppose you store a list of loaded DLLs within a process. Suppose that for
-  each of these DLLs you store the number of imported functions and each of the
-  imported functions name.
-
-  Imagine that a malicious DLL is injected into processes and its indicators are
-  that it only imports one function and that it is RegQueryValueEx. You would
-  write your indicator like this:
-
-  AndOperator(
-    Equal("ImportedDLLs.ImpFunctions.Name", "RegQueryValueEx"),
-    Equal("ImportedDLLs.NumImpFunctions", "1"))
-
-  Now imagine you have these two processes on a given system.
-
-  Process1
-  * __ImportedDlls
-
-    * __Name: "notevil.dll"
-
-      * __ImpFunctions
-
-        * __Name: "CreateFileA"
-
-      * __NumImpFunctions: 1
-
-    * __Name: "alsonotevil.dll"
-
-      * __ImpFunctions
-
-        * __Name: "RegQueryValueEx"
-        * __Name: "CreateFileA"
-
-      * __NumImpFunctions: 2
-
-  Process2
-  * __ImportedDlls
-
-    * __Name: "evil.dll"
-
-      * __ImpFunctions
-
-        * __Name: "RegQueryValueEx"
-
-      * __NumImpFunctions: 1
-
-  Both Process1 and Process2 match your query, as each of the indicators are
-  evaluated separately. While you wanted to express "find me processes that
-  have a DLL that has both one imported function and ReqQueryValueEx is in the
-  list of imported functions", your indicator actually means "find processes
-  that have at least a DLL with 1 imported functions and at least one DLL that
-  imports the ReqQueryValueEx function".
-
-  To write such an indicator you need to specify a context of ImportedDLLs for
-  these two clauses. Such that you convert your indicator to::
-
-      Context(
-        "ImportedDLLs",
-        AndOperator(
-          Equal("ImpFunctions.Name", "RegQueryValueEx"),
-          Equal("NumImpFunctions", "1")))
-
-  Context will execute the filter specified as the second parameter for each of
-  the objects under "ImportedDLLs", thus applying the condition per DLL, not per
-  object and returning the right result.
-  """
-
-  def __init__(self, arguments=None, **kwargs):
-    """Initializes a context operator.
-
-    Args:
-      arguments (Optional[object]): operands of the filter.
-
-    Raises:
-      InvalidNumberOfOperands: if the number of operands provided is not
-          supported.
-    """
-    if len(arguments) != 2:
-      raise errors.InvalidNumberOfOperands((
-          '{0:s} only supports 2 operands, provided were {1:d} '
-          'operands.').format(self.__class__.__name__, len(arguments)))
-
-    super(Context, self).__init__(arguments=arguments, **kwargs)
-    self.context, self.condition = self.args
-
-  def Matches(self, obj):
-    """Determines if the object matches the filter.
-
-    Args:
-      obj (object): object to compare against the filter.
-
-    Returns:
-      bool: True if the object matches the filter, False otherwise.
-    """
-    for object_list in self.value_expander.Expand(obj, self.context):
-      for sub_object in object_list:
-        if self.condition.Matches(sub_object):
-          return True
-    return False
-
-
-class ValueExpander(object):
-  """Value expander interface.
-
-  The value expander contains the logic to expand values in an object.
-  """
-
-  FIELD_SEPARATOR = '.'
-
-  def _GetAttributeName(self, path):
-    """Retrieves the attribute name to fetch given a path.
-
-    Args:
-      path (list[str]): object path, that contains the attribute names.
-
-    Returns:
-      str: attribute name.
-    """
-    return path[0]
-
-  @abc.abstractmethod
-  def _GetValue(self, obj, attr_name):
-    """Retrieves the value of a specific object attribute.
-
-    Args:
-      obj (object): object to retrieve the value from.
-      attr_name (str): name of the attribute to retrieve the value from.
-
-    Returns:
-      object: attribute value.
-    """
-
-  def _AtLeaf(self, attr_value):
-    """Retrieves the attribute value of a leaf node.
-
-    Yields:
-      object: attribute value.
-    """
-    yield attr_value
-
-  def _AtNonLeaf(self, attr_value, path):
-    """Retrieves the attribute value of a branch (non-leaf) node.
-
-    Yields:
-      object: attribute value.
-    """
-    try:
-      # Check first for iterables
-      # If it's a dictionary, we yield it
-      if isinstance(attr_value, dict):
-        yield attr_value
-      else:
-        # If it's an iterable, we recurse on each value.
-        for sub_obj in attr_value:
-          for value in self.Expand(sub_obj, path[1:]):
-            yield value
-    except TypeError:  # This is then not iterable, we recurse with the value
-      for value in self.Expand(attr_value, path[1:]):
-        yield value
-
-  def Expand(self, obj, path):
-    """Retrieves the attribute values from an object given an object path.
-
-    Given a path such as ["sub1", "sub2"] it returns all the values available
-    in obj.sub1.sub2 as a list. sub1 and sub2 must be data attributes or
-    properties.
-
-    If sub1 returns a list of objects, or a generator, Expand aggregates the
-    values for the remaining path for each of the objects, thus returning a
-    list of all the values under the given path for the input object.
-
-    Args:
-      obj (object): object that will be traversed.
-      path (list[str]): object path to traverse, that contains the attribute
-          names.
-
-    Yields:
-      object: attribute value.
-    """
-    if isinstance(path, py2to3.STRING_TYPES):
-      path = path.split(self.FIELD_SEPARATOR)
-
-    attr_name = self._GetAttributeName(path)
-    attr_value = self._GetValue(obj, attr_name)
-    if attr_value is None:
-      return
-
-    if len(path) == 1:
-      for value in self._AtLeaf(attr_value):
-        yield value
-    else:
-      for value in self._AtNonLeaf(attr_value, path):
-        yield value
-
-
-class AttributeValueExpander(ValueExpander):
-  """Value expander that expands based on object attribute names."""
-
-  def _GetValue(self, obj, attr_name):
-    """Retrieves the value of a specific object attribute.
-
-    Args:
-      obj (object): object to retrieve the value from.
-      attr_name (str): name of the attribute to retrieve the value from.
-
-    Returns:
-      object: attribute value or None if attribute is not available.
-    """
-    return getattr(obj, attr_name, None)
-
-
-class LowercaseAttributeValueExpander(AttributeValueExpander):
-  """Value expander that expands based on lower case object attribute names."""
-
-  def _GetAttributeName(self, path):
-    """Retrieves the attribute name to fetch given a path.
-
-    Args:
-      path (list[str]): object path, that contains the attribute names.
-
-    Returns:
-      str: attribute name.
-    """
-    return path[0].lower()
-
-
-class DictValueExpander(ValueExpander):
-  """Value expander that expands based on dictonary keys."""
-
-  def _GetValue(self, obj, attr_name):
-    """Retrieves the value of a specific object attribute.
-
-    Args:
-      obj (object): object to retrieve the value from.
-      attr_name (str): name of the attribute to retrieve the value from.
-
-    Returns:
-      object: attribute value or None if attribute is not available.
-    """
-    return obj.get(attr_name, None)
-
-
-# TODO: what is a "basic" expression?
-class BasicExpression(Expression):
-  """Basic expression.
-
-  Attribute:
-    bool_value (bool): boolean value that represents the result of
-        the operation.
-  """
-
-  def __init__(self):
-    """Initializes a basic expression."""
-    super(BasicExpression, self).__init__()
-    self.bool_value = True
-
-  def FlipBool(self):
-    """Negates the value of the bool_value attribute."""
-    self.bool_value = not self.bool_value
-
-  def Compile(self, filter_implementation):
-    """Given a filter implementation, compile this expression.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the binary expression.
-    """
-    arguments = [self.attribute]
-    op_str = self.operator.lower()
-    operator = filter_implementation.OPS.get(op_str, None)
-
-    if not operator:
-      raise errors.ParseError('Unknown operator {0:s} provided.'.format(
-          self.operator))
-
-    arguments.extend(self.args)
-    expander = filter_implementation.FILTERS['ValueExpander']
-    ops = operator(arguments=arguments, value_expander=expander)
-    if not self.bool_value:
-      if hasattr(ops, 'FlipBool'):
-        ops.FlipBool()
-
-    return ops
-
-
-class ContextExpression(Expression):
-  """Context operator expression."""
-
-  def __init__(self, attribute='', part=None):
-    """Initializes a context expression.
-
-    Args:
-      attribute (str): attribute.
-      part (str): expression part.
-    """
-    super(ContextExpression, self).__init__()
-    self.attribute = attribute
-    self.args = []
-    if part:
-      self.args.append(part)
-
-  def __str__(self):
-    """Retrieve a string representation of the expression.
-
-    Returns:
-      str: a string representation of the expression.
-    """
-    return 'Context({0:s} {1:s})'.format(
-        self.attribute, [str(x) for x in self.args])
-
-  def SetExpression(self, expression):
-    """Sets the expression.
-
-    Args:
-      expression (Expression): expression.
-
-    Raises:
-      ParseError: if expression is not of type Expression.
-    """
-    if not isinstance(expression, Expression):
-      raise errors.ParseError('Expected expression, got {0:s}.'.format(
-          type(expression)))
-
-    self.args = [expression]
-
-  def Compile(self, filter_implementation):
-    """Given a filter implementation, compile this expression.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the binary expression.
-    """
-    arguments = [self.attribute]
-    for argument in self.args:
-      arguments.append(argument.Compile(filter_implementation))
-    expander = filter_implementation.FILTERS['ValueExpander']
-    context_cls = filter_implementation.FILTERS['Context']
-    return context_cls(arguments=arguments,
-                       value_expander=expander)
-
-
-class PlasoExpression(BasicExpression):
-  """A Plaso specific expression."""
-
-  # A simple dictionary used to swap attributes so other names can be used
-  # to reference some core attributes (implementation specific).
-  swap_source = {
-      'date': 'timestamp',
-      'datetime': 'timestamp',
-      'time': 'timestamp',
-      'description_long': 'message',
-      'description': 'message',
-      'description_short': 'message_short',
-  }
-
-  def Compile(self, filter_implementation):
-    """Given a filter implementation, compile this expression.
-
-    Args:
-      filter_implementation (type): class of the filter object, which should
-          be a subclass of BaseFilterImplementation.
-
-    Returns:
-      object: filter object of the binary expression.
-    """
-    self.attribute = self.swap_source.get(self.attribute, self.attribute)
-    arguments = [self.attribute]
-    op_str = self.operator.lower()
-    operator = filter_implementation.OPS.get(op_str, None)
-
-    if not operator:
-      raise errors.ParseError('Unknown operator {0:s} provided.'.format(
-          self.operator))
-
-    # Plaso specific implementation - if we are comparing a timestamp
-    # to a value, we use our specific implementation that compares
-    # timestamps in a "human readable" format.
-    if self.attribute == 'timestamp':
-      args = []
-      for argument in self.args:
-        args.append(DateCompareObject(argument))
-      self.args = args
-
-    for argument in self.args:
-      if isinstance(argument, DateCompareObject):
-        if 'Less' in str(operator):
-          TimeRangeCache.SetUpperTimestamp(argument.data)
-        else:
-          TimeRangeCache.SetLowerTimestamp(argument.data)
-    arguments.extend(self.args)
-    expander = filter_implementation.FILTERS['ValueExpander']
-    ops = operator(arguments=arguments, value_expander=expander)
-    if not self.bool_value:
-      if hasattr(ops, 'FlipBool'):
-        ops.FlipBool()
-
-    return ops
-
-
 class EventFilterExpressionParser(object):
   """Event filter expression parser.
 
@@ -1273,9 +168,8 @@ class EventFilterExpressionParser(object):
   _ERROR_TOKEN = 'Error'
 
   # Classes of the expressions generated by the parser.
-  expression_cls = PlasoExpression
-  binary_expression_cls = BinaryExpression
-  context_cls = ContextExpression
+  expression_cls = expressions.EventExpression
+  binary_expression_cls = expressions.BinaryExpression
 
   tokens = [
       # Operators and related tokens
@@ -1349,8 +243,8 @@ class EventFilterExpressionParser(object):
     # Context can merge from item 0
     for i in range(len(self.stack)-1, 0, -1):
       item = self.stack[i-1]
-      if (isinstance(item, ContextExpression) and
-          isinstance(self.stack[i], Expression)):
+      if (isinstance(item, expressions.ContextExpression) and
+          isinstance(self.stack[i], expressions.Expression)):
         expression = self.stack[i]
         self.stack[i-1].SetExpression(expression)
         self.stack[i] = None
@@ -1365,10 +259,10 @@ class EventFilterExpressionParser(object):
     """
     for i in range(1, len(self.stack)-1):
       item = self.stack[i]
-      if (isinstance(item, BinaryExpression) and
+      if (isinstance(item, expressions.BinaryExpression) and
           item.operator.lower() == operator.lower() and
-          isinstance(self.stack[i-1], Expression) and
-          isinstance(self.stack[i+1], Expression)):
+          isinstance(self.stack[i-1], expressions.Expression) and
+          isinstance(self.stack[i+1], expressions.Expression)):
         lhs = self.stack[i-1]
         rhs = self.stack[i+1]
 
@@ -1382,7 +276,7 @@ class EventFilterExpressionParser(object):
     """Combines parenthesis."""
     for i in range(len(self.stack)-2):
       if (self.stack[i] == '(' and self.stack[i+2] == ')' and
-          isinstance(self.stack[i+1], Expression)):
+          isinstance(self.stack[i+1], expressions.Expression)):
         self.stack[i] = None
         self.stack[i+2] = None
 
@@ -1417,7 +311,8 @@ class EventFilterExpressionParser(object):
     Args:
       string (str): operator.
     """
-    self.stack.append(self.context_cls(string[1:]))
+    context_expression = expressions.ContextExpression(attribute=string[1:])
+    self.stack.append(context_expression)
 
   def Default(self, **kwarg):
     """Default callback handler."""
@@ -1645,7 +540,7 @@ class EventFilterExpressionParser(object):
       Expression: expression.
     """
     if not self.filter_string:
-      return IdentityExpression()
+      return expressions.IdentityExpression()
 
     self.Close()
     return self.Reduce()
@@ -1783,9 +678,6 @@ class EventFilterExpressionParser(object):
     self.current_expression.SetOperator(string)
 
 
-# Filter implementations.
-
-
 class BaseFilterImplementation(object):
   """Defines the base implementation of an object filter by its attributes.
 
@@ -1794,310 +686,22 @@ class BaseFilterImplementation(object):
   """
 
   OPS = {
-      'equals': Equals,
-      'is': Equals,
-      '==': Equals,
-      '!=': NotEquals,
-      'contains': Contains,
-      '>': Greater,
-      '>=': GreaterEqual,
-      '<': Less,
-      '<=': LessEqual,
-      'inset': InSet,
-      'regexp': Regexp,
-      'iregexp': RegexpInsensitive}
+      '==': filters.Equals,
+      '>': filters.Greater,
+      '>=': filters.GreaterEqual,
+      '<': filters.Less,
+      '<=': filters.LessEqual,
+      '!=': filters.NotEquals,
+      'contains': filters.Contains,
+      'equals': filters.Equals,
+      'inset': filters.InSet,
+      'iregexp': filters.RegexpInsensitive,
+      'is': filters.Equals,
+      'regexp': filters.Regexp}
 
   FILTERS = {
-      'ValueExpander': AttributeValueExpander,
-      'AndFilter': AndFilter,
-      'OrFilter': OrFilter,
-      'IdentityFilter': IdentityFilter,
-      'Context': Context}
-
-
-# pylint: disable=missing-type-doc,missing-return-type-doc,missing-return-doc
-
-
-class DictObject(object):
-  # There's a backslash in the class docstring, so as not to confuse Sphinx.
-  # pylint: disable=anomalous-backslash-in-string
-  """A simple object representing a dict object.
-
-  To filter against an object that is stored as a dictionary the dict
-  is converted into a simple object. Since keys can contain spaces
-  and/or other symbols they are stripped out to make filtering work
-  like it is another object.
-
-  Example dict::
-
-    {'A value': 234,
-     'this (my) key_': 'value',
-     'random': True,
-    }
-
-  This object would then allow access to object.thismykey that would access
-  the key 'this (my) key\_' inside the dict.
-  """
-
-  _STRIP_KEY_RE = re.compile(r'[ (){}+_=\-<>[\]]')
-
-  def __init__(self, dict_object):
-    """Initialize the object and build a secondary dict."""
-    # TODO: Move some of this code to a more value typed system.
-    self._dict_object = dict_object
-
-    self._dict_translated = {}
-    for key, value in dict_object.items():
-      key = self._StripKey(key)
-      self._dict_translated[key] = value
-
-  def _StripKey(self, key):
-    """Strips a key form symbols to use within a dictionary.
-
-    Args:
-      key (str): key to strip.
-
-    Returns:
-      str: stripped key.
-    """
-    return self._STRIP_KEY_RE.sub('', key.lower())
-
-  def __getattr__(self, attr):
-    """Return back entries from the dictionary."""
-    if attr in self._dict_object:
-      return self._dict_object.get(attr)
-
-    # Special case of getting all the key/value pairs.
-    if attr == '__all__':
-      ret = []
-      for key, value in self._dict_translated.items():
-        ret.append('{}:{}'.format(key, value))
-      return ' '.join(ret)
-
-    test = self._StripKey(attr)
-    if test in self._dict_translated:
-      return self._dict_translated.get(test)
-
-    return None
-
-
-class PlasoValueExpander(AttributeValueExpander):
-  """An expander that gives values based on object attribute names."""
-
-  def _GetMessage(self, event_object):
-    """Returns a properly formatted message string.
-
-    Args:
-      event_object: the event object (instance od EventObject).
-
-    Returns:
-      A formatted message string.
-    """
-    # TODO: move this somewhere where the mediator can be instantiated once.
-    formatter_mediator = formatters_mediator.FormatterMediator()
-
-    result = ''
-    try:
-      result, _ = formatters_manager.FormattersManager.GetMessageStrings(
-          formatter_mediator, event_object)
-    except KeyError as exception:
-      logging.warning(
-          'Unable to correctly assemble event with error: {0!s}'.format(
-              exception))
-
-    return result
-
-  def _GetSources(self, event_object):
-    """Returns properly formatted source strings.
-
-    Args:
-      event_object: the event object (instance od EventObject).
-    """
-    try:
-      # TODO: refactor to pass event and event_data as separate arguments.
-      source_short, source_long = (
-          formatters_manager.FormattersManager.GetSourceStrings(
-              event_object, event_object))
-    except KeyError as exception:
-      logging.warning(
-          'Unable to correctly assemble event with error: {0!s}'.format(
-              exception))
-
-    return source_short, source_long
-
-  def _GetValue(self, obj, attr_name):
-    ret = getattr(obj, attr_name, None)
-
-    if ret:
-      if isinstance(ret, dict):
-        ret = DictObject(ret)
-
-      if attr_name == 'tag':
-        return ret.labels
-
-      return ret
-
-    # Check if this is a message request and we have a regular EventObject.
-    if attr_name == 'message':
-      return self._GetMessage(obj)
-
-    # Check if this is a source_short request.
-    if attr_name in ('source', 'source_short'):
-      source_short, _ = self._GetSources(obj)
-      return source_short
-
-    # Check if this is a source_long request.
-    if attr_name in ('source_long', 'sourcetype'):
-      _, source_long = self._GetSources(obj)
-      return source_long
-
-    return None
-
-  def _GetAttributeName(self, path):
-    return path[0].lower()
-
-
-class PlasoAttributeFilterImplementation(BaseFilterImplementation):
-  """Does field name access on the lowercase version of names.
-
-  Useful to only access attributes and properties with Google's python naming
-  style.
-  """
-
-  FILTERS = {}
-  FILTERS.update(BaseFilterImplementation.FILTERS)
-  FILTERS.update({'ValueExpander': PlasoValueExpander})
-
-
-class DateCompareObject(object):
-  """A specific class created for date comparison.
-
-  This object takes a date representation, whether that is a direct integer
-  datetime object or a string presenting the date, and uses that for comparing
-  against timestamps stored in microseconds in in microseconds since
-  Jan 1, 1970 00:00:00 UTC.
-
-  This makes it possible to use regular comparison operators for date,
-  irrelevant of the format the date comes in, since plaso stores all timestamps
-  in the same format, which is an integer/long, it is a simple manner of
-  changing the input into the same format (int) and compare that.
-  """
-
-  def __init__(self, data):
-    """Take a date object and use that for comparison.
-
-    Args:
-      data: A string, datetime object or an integer containing the number
-            of micro seconds since January 1, 1970, 00:00:00 UTC.
-
-    Raises:
-      ValueError: if the date string is invalid.
-    """
-    if isinstance(data, py2to3.INTEGER_TYPES):
-      self.data = data
-      self.text = '{0:d}'.format(data)
-
-    elif isinstance(data, float):
-      self.data = py2to3.LONG_TYPE(data)
-      self.text = '{0:f}'.format(data)
-
-    elif isinstance(data, py2to3.STRING_TYPES):
-      if isinstance(data, py2to3.BYTES_TYPE):
-        self.text = data.decode('utf-8', errors='ignore')
-      else:
-        self.text = data
-
-      try:
-        self.data = timelib.Timestamp.FromTimeString(self.text)
-      except (ValueError, errors.TimestampError):
-        raise ValueError('Wrongly formatted date string: {0:s}'.format(
-            self.text))
-
-    elif isinstance(data, datetime.datetime):
-      posix_time = int(calendar.timegm(data.utctimetuple()))
-      self.data = (
-          posix_time * definitions.MICROSECONDS_PER_SECOND) + data.microsecond
-      self.text = '{0!s}'.format(data)
-
-    elif isinstance(data, DateCompareObject):
-      self.data = data.data
-      self.text = '{0!s}'.format(data)
-
-    else:
-      raise ValueError('Unsupported type: {0:s}.'.format(type(data)))
-
-  def __cmp__(self, x):
-    """A simple comparison operation."""
-    try:
-      x_date = DateCompareObject(x)
-
-      # The following implements a Python3 compatible:
-      # cmp(self.data, x_date.data)
-      return (self.data > x_date.data) - (self.data < x_date.data)
-    except ValueError:
-      return False
-
-  def __le__(self, x):
-    """Less or equal comparison."""
-    return self.data <= x
-
-  def __lt__(self, x):
-    """Less comparison"""
-    return self.data < x
-
-  def __ge__(self, x):
-    """Greater or equal comparison."""
-    return self.data >= x
-
-  def __gt__(self, x):
-    """Greater comparison."""
-    return self.data > x
-
-  def __eq__(self, x):
-    """Check if equal."""
-    return x == self.data
-
-  def __ne__(self, x):
-    """Check if not equal."""
-    return x != self.data
-
-  def __str__(self):
-    """Return a string representation of the object."""
-    return self.text
-
-
-class TimeRangeCache(object):
-  """A class that stores time ranges from filters."""
-
-  MAX_INT64 = 2**64-1
-
-  @classmethod
-  def SetLowerTimestamp(cls, timestamp):
-    """Sets the lower bound timestamp."""
-    if not hasattr(cls, '_lower'):
-      cls._lower = timestamp
-      return
-
-    if timestamp < cls._lower:
-      cls._lower = timestamp
-
-  @classmethod
-  def SetUpperTimestamp(cls, timestamp):
-    """Sets the upper bound timestamp."""
-    if not hasattr(cls, '_upper'):
-      cls._upper = timestamp
-      return
-
-    if timestamp > cls._upper:
-      cls._upper = timestamp
-
-  @classmethod
-  def GetTimeRange(cls):
-    """Return the first and last timestamp of filter range."""
-    first = getattr(cls, '_lower', 0)
-    last = getattr(cls, '_upper', cls.MAX_INT64)
-
-    if first < last:
-      return first, last
-
-    return last, first
+      'AndFilter': filters.AndFilter,
+      'Context': filters.Context,
+      'IdentityFilter': filters.IdentityFilter,
+      'OrFilter': filters.OrFilter,
+      'ValueExpander': value_expanders.PlasoValueExpander}
