+++ tests/filters/filters.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Tests for the event filter expression parser filter classes."""
+
+from __future__ import unicode_literals
+
+import unittest
+
+
+# TODO: add tests
+
+
+if __name__ == "__main__":
+  unittest.main()
